requirements.txt
asgiref==3.9.1
crispy-tailwind==1.0.3
Django==5.2.5
django-crispy-forms==2.4
django-environ==0.12.0
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.2
